[refactor] 앞선 테스트 설정 파일 구조 및 수정으로 인한 테스트 코드 수정
