+++ src/test/java/kr/hhplus/be/server/integration/concurrency/OrderCreateAndOrderCancelTest.java
@@ -1,6 +1,5 @@
-package kr.hhplus.be.server.concurrency;
+package kr.hhplus.be.server.integration.concurrency;
 
-import kr.hhplus.be.server.ServerApplication;
 import kr.hhplus.be.server.TestContainersConfiguration;
 import kr.hhplus.be.server.application.coupon.service.CouponService;
 import kr.hhplus.be.server.application.order.dto.OrderRequest;
@@ -12,12 +11,9 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.aop.support.AopUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.jdbc.Sql;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
@@ -30,13 +26,9 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-@SpringBootTest(classes = {ServerApplication.class, TestContainersConfiguration.class})
-@ActiveProfiles("test")
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE) // 테스트컨테이너에서 외부 DB 사용하도록 함
 @Sql(scripts = "/concurrencyOrder.sql",executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS) //테스트 실행 시 해당 .sql 파일내의 쿼리 실행 -> 테이블 생성 후 실행됨
-@Sql(scripts = "/delete.sql",executionPhase = Sql.ExecutionPhase.AFTER_TEST_CLASS) //이 클래스 테스트 종료 시 데이터 클랜징
 @EnableAspectJAutoProxy // AOP 활성화
-public class OrderCreateAndOrderCancelTest {
+public class OrderCreateAndOrderCancelTest extends TestContainersConfiguration {
 
     @Autowired
     OrderFacadeService orderFacadeService;
@@ -54,6 +46,7 @@ public class OrderCreateAndOrderCancelTest {
     }
 
     @Test
+    @Transactional
     @DisplayName("주문 생성 vs 주문 취소 동시성 테스트")
     void orderCreateAndOrderCancel() throws Exception {
         //Given
@@ -160,6 +153,7 @@ public class OrderCreateAndOrderCancelTest {
     }
 
     @Test
+    @Transactional
     @DisplayName("주문 생성 vs 주문 취소 동시성 테스트 - 재고 반복 차감")
     void orderCreateAndOrderCancel2() throws Exception {
         //Given
@@ -314,6 +308,7 @@ public class OrderCreateAndOrderCancelTest {
 
 
     @Test
+    @Transactional
     @DisplayName("주문 생성 vs 주문 취소 동시성 테스트 - 재고 반복 복구")
     void orderCreateAndOrderCancel3() throws Exception {
         //Given
