+++ src/test/java/kr/hhplus/be/server/integration/concurrency/PaymentAndOrderCancelTest.java
@@ -1,21 +1,23 @@
-package kr.hhplus.be.server.concurrency;
+package kr.hhplus.be.server.integration.concurrency;
 
-import kr.hhplus.be.server.ServerApplication;
 import kr.hhplus.be.server.TestContainersConfiguration;
-import kr.hhplus.be.server.application.balance.dto.BalanceRequest;
-import kr.hhplus.be.server.application.balance.dto.BalanceResponse;
 import kr.hhplus.be.server.application.balance.service.BalanceService;
+import kr.hhplus.be.server.application.coupon.service.CouponService;
+import kr.hhplus.be.server.application.order.dto.OrderRequest;
+import kr.hhplus.be.server.application.order.dto.OrderResponse;
+import kr.hhplus.be.server.application.order.facade.OrderFacadeService;
+import kr.hhplus.be.server.application.order.repository.OrderRepository;
 import kr.hhplus.be.server.application.payment.dto.PaymentRequest;
+import kr.hhplus.be.server.application.payment.dto.PaymentResponse;
 import kr.hhplus.be.server.application.payment.facade.PaymentFacadeService;
+import kr.hhplus.be.server.application.product.service.ProductService;
+import kr.hhplus.be.server.domain.order.Order;
+import org.apache.commons.lang3.StringUtils;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.aop.support.AopUtils;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.EnableAspectJAutoProxy;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.jdbc.Sql;
 
 import java.time.LocalDateTime;
@@ -27,38 +29,46 @@ import java.util.concurrent.Executors;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
-@SpringBootTest(classes = {ServerApplication.class, TestContainersConfiguration.class})
-@ActiveProfiles("test")
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE) // 테스트컨테이너에서 외부 DB 사용하도록 함 //application-test.yml 읽어오도록 함
-@Sql(scripts = "/concurrencyPaymentAndBalance.sql",executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS) //테스트 실행 시 해당 .sql 파일내의 쿼리 실행 -> 테이블 생성 후 실행됨
-@Sql(scripts = "/delete.sql",executionPhase = Sql.ExecutionPhase.AFTER_TEST_CLASS) //이 클래스 테스트 종료 시 데이터 클랜징
+@Sql(scripts = "/concurrencyPaymentAndOrderCancel.sql",executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS) //테스트 실행 시 해당 .sql 파일내의 쿼리 실행 -> 테이블 생성 후 실행됨
 @EnableAspectJAutoProxy // AOP 활성화
-public class PaymentAndBalanceChargeTest {
+public class PaymentAndOrderCancelTest extends TestContainersConfiguration {
 
     @Autowired
     PaymentFacadeService paymentFacadeService;
 
     @Autowired
+    OrderFacadeService orderFacadeService;
+
+    @Autowired
+    OrderRepository orderRepository;
+
+    @Autowired
     BalanceService balanceService;
+
+    @Autowired
+    ProductService productService;
+
+    @Autowired
+    CouponService couponService;
+
 
     //프록시 등록 상태 확인
     @Test
     void checkProxy() {
-        System.out.println("Proxy balanceService? " + AopUtils.isAopProxy(balanceService));
+        System.out.println("Proxy orderFacadeService? " + AopUtils.isAopProxy(orderFacadeService));
         System.out.println("Proxy paymentFacadeService? " + AopUtils.isAopProxy(paymentFacadeService));
     }
 
     @Test
-    @DisplayName("결제 vs 잔액 충전 동시성 테스트")
-    void paymentAndBalanceCharge() throws Exception {
+    @DisplayName("결제 vs 주문 취소 동시성 테스트")
+    void paymentAndOrderCancel() throws Exception {
         //Given
         //파라미터 셋팅
         long userId = 1L;
         long orderId = 3L;
         PaymentRequest.Create createPayment = new PaymentRequest.Create(userId,orderId);
 
-        long chargeAmount = 10_000L;
-        BalanceRequest chargeBalance = new BalanceRequest(userId,chargeAmount);
+        OrderRequest.OrderCancel cancelOrder = new OrderRequest.OrderCancel(userId,orderId);
 
         //동시성 테스트 준비
         int threadCount = 2;
@@ -79,7 +89,8 @@ public class PaymentAndBalanceChargeTest {
                 startBarrier.await();
 
                 //결제
-                paymentFacadeService.createPayment(createPayment);
+                PaymentResponse.Create result = paymentFacadeService.createPayment(createPayment);
+                System.out.println("결제 완료 후 주문 상태: "+result.order().orderStatus());
             } catch (Exception e) {
                 System.err.println("스레드 실행 중 오류: " + e.getMessage());
             } finally {
@@ -93,8 +104,9 @@ public class PaymentAndBalanceChargeTest {
                 //모든 스레드가 Barrier에 도착할 때까지 대기
                 startBarrier.await();
 
-                //잔액 충전
-                balanceService.charge(chargeBalance);
+                //주문 취소
+                OrderResponse.OrderDTO result = orderFacadeService.cancelOrder(cancelOrder);
+                System.out.println("주문 취소 후 주문 상태: "+result.orderStatus());
             } catch (Exception e) {
                 System.err.println("스레드 실행 중 오류: " + e.getMessage());
             } finally {
@@ -119,11 +131,20 @@ public class PaymentAndBalanceChargeTest {
         System.out.println("모든 스레드 실행 종료: " + endTime.format(formatter));
 
         //Then
-        BalanceResponse result = balanceService.selectBalanceByUserId(userId);
-        System.out.println("사용자 잔액: "+result.balance());
-
-        assertEquals(91_000L,result.balance());
+        Order result = orderRepository.findByOrderId(orderId);
+        System.out.println("주문 상태: "+result.getOrderStatus());
+        //주문 취소 후 결제 진행된 케이스
+        if(StringUtils.equals(result.getOrderStatus(),"payment_completed")){
+            assertEquals("payment_completed",result.getOrderStatus());
+            assertEquals(30L,productService.selectProductOptionByProductIdAndProductOptionId(1L,1L).getStockQuantity());
+            assertEquals(30L,productService.selectProductOptionByProductIdAndProductOptionId(1L,2L).getStockQuantity());
+            assertEquals(30L,productService.selectProductOptionByProductIdAndProductOptionId(1L,3L).getStockQuantity());
+            assertEquals(30L,productService.selectProductOptionByProductIdAndProductOptionId(1L,4L).getStockQuantity());
+            assertEquals(30L,productService.selectProductOptionByProductIdAndProductOptionId(1L,5L).getStockQuantity());
+            assertEquals("N",couponService.selectCouponByCouponIdAndUserId(3L,userId).getUseYn());
+        }else{//결제 완료 후 주문 취소된 케이스
+            assertEquals("cancel_order",result.getOrderStatus());
+            assertEquals(400_000L,balanceService.selectBalanceByUserId(userId).balance());
+        }
     }
-
-
 }
