+++ src/test/java/kr/hhplus/be/server/integration/persistence/coupon/CouponIssuedInfoJpaRepositoryTest.java
@@ -1,18 +1,12 @@
 package kr.hhplus.be.server.integration.persistence.coupon;
 
-import kr.hhplus.be.server.ServerApplication;
 import kr.hhplus.be.server.TestContainersConfiguration;
 import kr.hhplus.be.server.domain.coupon.CouponIssuedInfo;
 import kr.hhplus.be.server.persistence.coupon.CouponIssuedInfoAdapter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.test.annotation.Commit;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.jdbc.Sql;
 import org.springframework.transaction.annotation.Transactional;
 
@@ -21,19 +15,14 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-@SpringBootTest(classes = {ServerApplication.class, TestContainersConfiguration.class})
-@ActiveProfiles("test")
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE) // 테스트컨테이너에서 외부 DB 사용하도록 함
 @Sql(scripts = "/couponIssuedInfo.sql",executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS) //테스트 실행 시 해당 .sql 파일내의 쿼리 실행 -> 테이블 생성 후 실행됨
-@Sql(scripts = "/delete.sql",executionPhase = Sql.ExecutionPhase.AFTER_TEST_CLASS) //이 클래스 테스트 종료 시 데이터 클랜징
 @ComponentScan(basePackageClasses = CouponIssuedInfoAdapter.class)//@Component 사용 중인 Adapter 클래스 읽어오기 위함
-class CouponIssuedInfoJpaRepositoryTest {
+class CouponIssuedInfoJpaRepositoryTest extends TestContainersConfiguration {
 
     @Autowired
     CouponIssuedInfoAdapter couponIssuedInfoAdapter;
 
     @Test
-    @Transactional
     @DisplayName("쿠폰 ID, 사용자 ID로 보유 쿠폰 발급 정보 조회- findByCouponIdAndUserId()")
     void findByCouponIdAndUserId(){
         System.out.println("findByCouponIdAndUserId 쿼리");
@@ -108,7 +97,6 @@ class CouponIssuedInfoJpaRepositoryTest {
     }
 
     @Test
-    @Transactional
     @DisplayName("사용자 ID로 소유 쿠폰 조회 - findByUserId()")
     void findByUserId(){
         System.out.println("findByUserId 쿼리");
