+++ src/test/java/kr/hhplus/be/server/integration/persistence/product/ProductStatisticsAdapterTest.java
@@ -1,7 +1,6 @@
 package kr.hhplus.be.server.integration.persistence.product;
 
 import com.querydsl.core.Tuple;
-import kr.hhplus.be.server.ServerApplication;
 import kr.hhplus.be.server.TestContainersConfiguration;
 import kr.hhplus.be.server.application.order.dto.OrderProductSummary;
 import kr.hhplus.be.server.domain.order.QOrderProduct;
@@ -10,31 +9,21 @@ import kr.hhplus.be.server.persistence.product.ProductStatisticsAdapter;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.jdbc.AutoConfigureTestDatabase;
-import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.context.annotation.ComponentScan;
-import org.springframework.test.context.ActiveProfiles;
-import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.jdbc.Sql;
-import org.springframework.transaction.annotation.Transactional;
 
 import java.util.ArrayList;
 import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
-@SpringBootTest(classes = {ServerApplication.class, TestContainersConfiguration.class})
-@ActiveProfiles("test")
-@AutoConfigureTestDatabase(replace = AutoConfigureTestDatabase.Replace.NONE) // 테스트컨테이너에서 외부 DB 사용하도록 함
 @Sql(scripts = "/productStatistics.sql",executionPhase = Sql.ExecutionPhase.BEFORE_TEST_CLASS) //테스트 실행 시 해당 .sql 파일내의 쿼리 실행 -> 테이블 생성 후 실행됨
-@Sql(scripts = "/delete.sql",executionPhase = Sql.ExecutionPhase.AFTER_TEST_CLASS) //테스트 실행 시 해당 .sql 파일내의 쿼리 실행 -> 테이블 생성 후 실행됨
 @ComponentScan(basePackageClasses = ProductStatisticsAdapter.class)//@Component 사용 중인 Adapter 클래스 읽어오기 위함
-class ProductStatisticsAdapterTest {
+class ProductStatisticsAdapterTest extends TestContainersConfiguration {
 
     @Autowired
     ProductStatisticsAdapter productStatisticsAdapter;
 
     @Test
-    @Transactional
     @DisplayName("지난3일 인기상품 TOP5 조회 - QureyDSL")
     void queryDsl() {
         System.out.println("queryDSL - 쿼리");
